min coupling unplag_core
